Fix lint errors in bin/server.ts
- Add type annotations to top-level version-check block constants
- Extract logger setup into bin/logger.ts to fit under max-lines=500
